+++ CHANGELOG.md
@@ -13,9 +13,13 @@
   resulting lobster file will now contain "Unset" as status information,
   too, instead of `Null`.
 
+* The `lobster-python` tool now adds the counter logic to the function
+  identifier. This supports situations where different functions have
+  the same name.
+
 ### 0.9.17
 
-* The `lobster-python` tool now adds the counter logic to the function
+* The `lobster-python` tool now adds the line number to the function
   identifier. This supports situations where different functions have
   the same name